+++ internal/engine/wazevo/e2e_test.go
@@ -1049,7 +1049,7 @@ func TestE2E_stores(t *testing.T) {
 
 	f := inst.ExportedFunction(testcases.ExportedFunctionName)
 
-	mem, ok := inst.Memory().Read(0, uint64(wasm.MemoryPageSize))
+	mem, ok := inst.Memory().Read(0, wasm.MemoryPageSize)
 	require.True(t, ok)
 	for _, tc := range []struct {
 		i32 uint32
@@ -1144,7 +1144,7 @@ func TestE2E_reexported_memory(t *testing.T) {
 	mem := m1Inst.Memory()
 	require.Equal(t, mem, m3Inst.Memory())
 	require.Equal(t, mem, m2Inst.Memory())
-	require.Equal(t, uint64(11), mem.Size()/65536)
+	require.Equal(t, uint32(11), mem.Size()/65536)
 }
 
 func TestStackUnwind_panic_in_host(t *testing.T) {
